Add: 30 runs of robot using two cameras, and a new script to convert bag to .csv data

+++ localization_analysis/localization_analysis/localization_logger.py
@@ -58,7 +58,7 @@ class LocalizationLogger(Node):
     def __init__(self):
         super().__init__('localization_logger')
 
-        self.declare_parameter('camera_mode', 'single')
+        self.declare_parameter('camera_mode', 'double')
 
         run_dir = _next_run_dir()
         default_output = os.path.join(run_dir, 'localization_log.csv')
